+++ rs/sns/integration_tests/src/lib.rs
@@ -54,3 +54,6 @@ mod manage_dapp_canister_settings;
 
 #[cfg(test)]
 mod http_request;
+
+#[cfg(test)]
+mod timers;